# docs/swagger.yaml
definitions:
...
  controllers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  user.NewUser:
    properties:
      age:
        example: 1
        type: integer
      email:
        example: mail@mail.com
        type: string
      password:
        example: Password123
        type: string
      role_id:
        type: string
      username:
        example: someUser
        type: string
    required:
    - age
    - email
    - password
    - role_id
    - username
    type: object
  user.ResponseUser:
    properties:
      age:
        example: 1
        type: integer
      createdAt:
        example: "2021-02-24 20:19:39"
        type: string
      deleted_at:
        example: "null"
        type: string
      email:
        example: some@mail.com
        type: string
      id:
        example: 1099
        type: integer
      updatedAt:
        example: "2021-02-24 20:19:39"
        type: string
      user:
        example: BossonH
        type: string
    required:
    - age
    type: object
...
info:
  contact: {}
paths:
...
  /user:
    get:
      description: Get all Users on the system
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.ResponseUser'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all Users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create new user on the system
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.NewUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ResponseUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create New UserName
      tags:
      - user
  /user/{user_id}:
    get:
      description: Get Users by ID on the system
      parameters:
      - description: id of user
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get users by ID
      tags:
      - user
swagger: "2.0"